ViewPrivateComment Permissions
Added the new permission ViewPrivateComment. Gave this permission to the
Pre-defined Roles: Project Administrators, Quality Assurance, and
Developer.

Wasn't sure about where to put new sql changes so created
1.6.330.0.SqlDataProvider.sql (as well as adding to
BugNet.Data.SqlDataProvider.sql), as last available version was
1.6.329.0.

Also added GB as a Language (for use by those of us in UK, Australia and
NZ) to get the dates in the expected format.