Enable use of local avatar
File can be uploaded via User Profile and will be saved in the
Images\Users folder as two jpgs of specific size. If gravatars are
allowed, and selected, these will be disabled so can be used if later
decide against using a gravatar.

Not happy with how this was done as have persistence in the presentation
layer. Needs to be in data layer, and also probably better to save/get
the url (of either the local avatar, or the gravatar) in the database,
in the BugNet_UserProfiles table.

[not sure why can't commit new Images\Users folder]