Completed finer granularity of notification prefs
NotificationOfChanges as enum in Global Enums
NotificationOfChanges as new column in BugNet_UserProfiles
(nvarchar(255)) to take space separated list of enum values. Default is
all ("1 2 3 4 5") to keep existing behaviour
alter table in latest.sql
modified table create in schema .sql
IssueNotificationManager.cs
SendIssueAddNotification - called when adding a new issue to a project,
so only send to user who has
NotificationOfChanges.IssueAddedToMonitoredProject set
SendIssueNotifications - overloaded
(int) - unused and would error if was (as assumes there is
IssueHistoryChanges.IssueHistory data attached) so set to obsolete
(int, IEnumerable<IssueHistory>) - called when anything changed in the
issue's columns (including adding a Time Entry as that increased the
hours logged).
If the user has NotificationOfChanges.IssueStatusChanged set and there
is a history change on the "Status" field, then sends a notification.
Also if the user has NotificationOfChanges.IssueOtherColumnChanged set
and there is any history change on a field other than "Status", then
sends a notification
SendNewAssigneeNotification - called when user is assigned to project,
so only send if user has NotificationOfChanges.IssueAssignedToReceiver
set
SendNewIssueCommentNotification - called when a comment is added to an
issue, so only send if user has NotificationOfChanges.IssueCommentAdded
set
UserProfile.aspx. Set these values in the Notifications section. --
still working on getting the values there from the enums --
WebProfile.cs (in BLL) stores the NotificationOfChanges string
(space-separated list of enum values) for storing to the database which
is handled automatically. Had to set this new property value in
web.config aswell as specifying database type.