Issue Subject formatted text uses named parameters
IssueUpdatedSubject (in Global Notifications.resx) can now use any of
the following named parameters:
* IssueId
* IssueTitle
* DisplayName
* FieldName
* NewValue
* OldValue
where the FieldName, NewValue and OldValue are from the IssueHistory
selected to have the highest priority (based on FieldChanged). At the
moment this priority order is hard-coded as: "Title", "Status",
"Resolution", "Priority", "Description"

Created a FormatContentWith function in CultureContent to be used
instead of the FormatContent. Based on code from:
http://james.newtonking.com/archive/2008/03/29/formatwith-2-0-string-formatting-with-named-variables

Would need to have, perhaps, a different subject if the value was long
(such as for Description...perhaps just truncate if using?), or if the
Title (as often may want to say Issue "OldValue" (id) title has changed
to "NewValue", rather than Issue "Title" (id) FieldName has changed to
"NewValue"